deps: Bump sqlfluff to 2.1.4, DO NOT remove pilo_star
pilo_star was introduced because of this bug that we explained and
reported.

I thought it could be removed and that we could use the original
dbt_utils.star again, but no :'(

It does not work with GROUP BY statements because on CI it would
generate a `GROUP BY *` statement, which is illegal.

Our version generates a `GROUP BY id` statement.

So.... let's stick to it cause it will not be fixed.