SQLFluff: Ignore a unsolvable (for now) rule violation
This valid error is only seen with `CI=1`, ignoring it as we can't
really solve it right now.

When the `pilo_star` macro is replaced by `id` and not `*` or the
expanded list of the table's fields then we get `candidatures.id` two
times, one via the macro and another directly in the file.