Create DAG populate_metabase_fluxiae
Each week a DAG downloads the latest FluxIAE import from a bucket shared with les-emplois, decrypts and processes the data contained within to Metabase. If new data does not exist, the task fails loudly, but this can be overridden by an administrator.

The commit adds a new dependency (psycopg3), and some new secrets that are mostly configured by the Airtable admin but the hash salt is configured by environment variable on the instance directly.