dags: Remove `dbt clean` tasks
Calling `dbt clean` while other dbt commands runs will cause them to
fails because of a missing dependency, so before trying to schedule them
manually we are going to check if this is needed.

`dbt deps` use a lock file so dependencies will be upgrade if needed.

And as per the documentation it should be OK, especially since we run it
inside a docker container in production:
> It helps by removing unnecessary files or directories generated during
> the execution of other dbt commands, ensuring a clean state for the
> project.